band steering: fix auto mode work.

+++ linux-3.4.x/drivers/net/wireless/ralink/rtpci/mt7610/include/band_steering_def.h
@@ -134,7 +134,7 @@ enum BND_STRG_CONDITION_CHECK_FLAGS {
 typedef struct _BNDSTRG_MSG{
 	UINT8	 Action;
 	UINT8	 ReturnCode;
-	UINT8	 TalbeIndex;
+	UINT32	 TalbeIndex;
 	BOOLEAN OnOff;
 	UINT8	Band;
 	BOOLEAN b2GInfReady;
